requirements.txt
asgiref==3.11.1
defusedxml==0.7.1
dj-database-url==3.1.0
Django==6.0.2
django-filter==25.2
django-simple-history==3.11.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.61.1

fpdf2==2.8.5
gunicorn==25.0.3
Markdown==3.10.2
numpy==2.4.2
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
qrcode==8.2
six==1.17.0
sqlparse==0.5.5
whitenoise==6.11.0
xlrd==2.0.2
matplotlib==3.10.0
